price scraping adjustment
The scraper was initially tested only on UK based pages that used "pounds" as currency. However, when providing a page with a different currency the price defaults to -1.00 because currency indicators like €, $, ₺ would not be removed from the string, which prevented it from successfully being converted to a float type.